Initial vars configuration
I want to make it easier to use in other distros, like debian
on my raspberry pi. I just found that some packages differ in
name, or doesn't exist on debian, and other things that I
would like to do on my debian, but don't want it on my
fedora.
By mistake I also added the system-linux.yml that is just
fixing an ansible warning, but I'm too lazy to remove from
the commit